+++ src/connectedvmware/azext_connectedvmware/vendored_sdks/operations/_vcenters_operations.py
@@ -32,7 +32,7 @@ class VCentersOperations(object):
     instantiates it for you and attaches it as an attribute.
 
     :ivar models: Alias to model classes used in this operation group.
-    :type models: ~azure_arc_vmware_management_service_api.models
+    :type models: ~azure.mgmt.connectedvmware.models
     :param client: Client for service requests.
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
@@ -126,7 +126,7 @@ class VCentersOperations(object):
         :param vcenter_name: Name of the vCenter.
         :type vcenter_name: str
         :param body: Request payload.
-        :type body: ~azure_arc_vmware_management_service_api.models.VCenter
+        :type body: ~azure.mgmt.connectedvmware.models.VCenter
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
         :keyword polling: By default, your polling method will be ARMPolling.
@@ -134,7 +134,7 @@ class VCentersOperations(object):
         :paramtype polling: bool or ~azure.core.polling.PollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
         :return: An instance of LROPoller that returns either VCenter or the result of cls(response)
-        :rtype: ~azure.core.polling.LROPoller[~azure_arc_vmware_management_service_api.models.VCenter]
+        :rtype: ~azure.core.polling.LROPoller[~azure.mgmt.connectedvmware.models.VCenter]
         :raises ~azure.core.exceptions.HttpResponseError:
         """
         polling = kwargs.pop('polling', True)  # type: Union[bool, PollingMethod]
@@ -200,7 +200,7 @@ class VCentersOperations(object):
         :type vcenter_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: VCenter, or the result of cls(response)
-        :rtype: ~azure_arc_vmware_management_service_api.models.VCenter
+        :rtype: ~azure.mgmt.connectedvmware.models.VCenter
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.VCenter"]
@@ -262,10 +262,10 @@ class VCentersOperations(object):
         :param vcenter_name: Name of the vCenter.
         :type vcenter_name: str
         :param body: Resource properties to update.
-        :type body: ~azure_arc_vmware_management_service_api.models.ResourcePatch
+        :type body: ~azure.mgmt.connectedvmware.models.ResourcePatch
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: VCenter, or the result of cls(response)
-        :rtype: ~azure_arc_vmware_management_service_api.models.VCenter
+        :rtype: ~azure.mgmt.connectedvmware.models.VCenter
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.VCenter"]
@@ -449,7 +449,7 @@ class VCentersOperations(object):
 
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either VCentersList or the result of cls(response)
-        :rtype: ~azure.core.paging.ItemPaged[~azure_arc_vmware_management_service_api.models.VCentersList]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.mgmt.connectedvmware.models.VCentersList]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.VCentersList"]
@@ -522,7 +522,7 @@ class VCentersOperations(object):
         :type resource_group_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either VCentersList or the result of cls(response)
-        :rtype: ~azure.core.paging.ItemPaged[~azure_arc_vmware_management_service_api.models.VCentersList]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.mgmt.connectedvmware.models.VCentersList]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.VCentersList"]