+++ src/vmware/HISTORY.md
@@ -16,6 +16,9 @@
 - Deprecate `az vmware private-cloud listadmincredentials`
 - Deprecate `az vmware private-cloud addidentitysource`
 - Deprecate `az vmware private-cloud deleteidentitysource`
+- [BREAKING CHANGE] Add confirmation prompt to `az vmware private-cloud rotate-vcenter-password`
+- [BREAKING CHANGE] Add confirmation prompt to `az vmware private-cloud rotate-nsxt-password`
+- [BREAKING CHANGE] Add confirmation prompt to all delete commands
 
 ## 3.2.0 (2021-10)