+++ src/vmware/HISTORY.md
@@ -13,6 +13,7 @@
 - Add `az vmware private-cloud list-admin-credentials`
 - Add `az vmware private-cloud add-identity-source`
 - Add `az vmware private-cloud delete-identity-source`
+- Add `az vmware private-cloud identity` command group
 - Deprecate `az vmware private-cloud listadmincredentials`
 - Deprecate `az vmware private-cloud addidentitysource`
 - Deprecate `az vmware private-cloud deleteidentitysource`